style/refactor: make quantity buttons use button component + style cart items
