+++ src/app/cart/components/CartItems/CartItems.module.css
@@ -1,0 +1,19 @@
+.cartItems {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.cartItems__title {
+  border: 1px solid var(--clr-neutral-300);
+  padding: 1rem;
+  border-radius: 1rem;
+}
+
+.cartItems__items {
+  display: flex;
+  flex-direction: column;
+  border: 1px solid var(--clr-neutral-300);
+  border-radius: 1rem;
+  padding-inline: 1rem;
+}
