+++ src/app/cart/components/RemoveButton/styles.module.css
@@ -1,5 +1,5 @@
 .removeButton {
-  background-color: var(--clr-neutral-200);
+  background-color: var(--clr-neutral-200) !important;
   width: calc(var(--button-height) * 0.5);
   height: calc(var(--button-height) * 0.5);
   min-width: calc(var(--button-height) * 0.5);
@@ -7,6 +7,7 @@
   display: flex;
   align-items: center;
   justify-content: center;
+  border-radius: 1rem;
 }
 
 .removeButton svg {
