+++ src/app/cart/styles.module.css
@@ -10,34 +10,6 @@
   gap: 1rem;
 }
 
-.cartItems {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.cartItems__title {
-  border: 1px solid var(--clr-neutral-300);
-  padding: 1rem;
-  border-radius: 1rem;
-}
-
-.cartItems__items {
-  display: flex;
-  flex-direction: column;
-  border: 1px solid var(--clr-neutral-300);
-  border-radius: 1rem;
-  padding-inline: 1rem;
-}
-
-.cartItem {
-  border-bottom: 1px solid var(--clr-neutral-300);
-}
-
-.cartItem:last-child {
-  border-bottom: none;
-}
-
 .cartSummary {
   position: sticky;
   top: 1rem;
